CalendarEvent team_idx 타입 수정
- CalendarEvent Entity: team_idx를 Long에서 Integer로 변경 (DB 컬럼 타입 일치)
- CalendarEventDto: team_idx 타입 Integer로 변경
- CalendarEventService: team_idx 타입 변환 처리 추가

+++ src/main/java/com/pinecni/erp/api/calendar/dto/CalendarEventDto.java
@@ -66,7 +66,7 @@ public class CalendarEventDto {
     private String status;
 
     // 팀 관련 정보
-    private Long teamIdx;
+    private Integer teamIdx;
     private String teamName;
     private String teamColor;
